Bugfixes from testing with ESIF dataset (show drilldown chart for some rules instead of grouped, use long instead of int, do not generate condition in task pmml if there is no condition...)

+++ Recommender.Business/DTO/DrilldownGraphDto.cs
@@ -11,7 +11,7 @@ namespace Recommender.Business.DTO
     {
         public override double GetValue()
         {
-            return XAxisLeaves.Aggregate(0, (current, leaf) => (int) (current + ((DrilldownGraphXAxisLeafDto) leaf).Value));
+            return XAxisLeaves.Aggregate(0L, (current, leaf) => (long) (current + ((DrilldownGraphXAxisLeafDto) leaf).Value));
         }
     }
